Created employee portal. This currently links to the manager view if the current person that logged in classifies as a "Manager". If not they only have the other two choices: take order or log out. Logout functionality works and take order is being worked on by another teammate.
