+++ frontend/src/app/ManagerScreen/components/Navbar.js
@@ -5,12 +5,12 @@ const Navbar = () => {
     <nav className="bg-[var(--navbar-bg)] text-[var(--navbar-text)] h-16 flex items-center justify-between px-8">
       <ul className="flex space-x-8">
         <li>
-          <Link href="/" className="navbarLinks">Home</Link>
+          <Link href="/employeeHome" className="navbarLinks">Home</Link>
         </li>
       </ul>
       <ul className="flex space-x-8">
         <li>
-          <Link href="/ManagerScreen/personnel" className="navbarLinks">Personnel</Link>
+          <Link href="/ManagerScreen" className="navbarLinks">Personnel</Link>
         </li>
         <li>
           <Link href="/ManagerScreen/menu" className="navbarLinks">Menu</Link>
